Rename routing to transition and update logic for seamless
