+++ client/src/index.jsx
@@ -10,7 +10,7 @@ import LandingContainer from './landing/LandingContainer'
 import {Provider} from 'react-redux'
 import RootComponent from './root/Root'
 import { fetchConfig } from './config/ConfigActions'
-import { setOperation, instantiateAppState } from './routing/RoutingActions'
+import { setOperation, instantiateAppState } from './transition/TransitionActions'
 import '../style/style'
 import './page.css'
 import store from './store'
